requirements.txt
pandas
sentence-transformers
langchain
langchain_core
pinecone-client
python-dotenv
langchain_groq
streamlit
openai 
langchain_openai
langchain-ollama